+++ NEWS.md
@@ -1,6 +1,10 @@
+# parkR 0.2.1
+
+Added basic features, transforms, and metadata for WJD, Omnibook and Essen Folks Song Collection
+
 # parkR 0.2.0
 
-Added harmonic analysis, added extended irb data.
+Added harmonic analysis, added extended iRealBook (irb) data.
 
 # parkR 0.1.1